+++ BangBangGokGok_springBoot/src/main/java/com/kosmo/project/review/reply/model/ReviewReplyVO.java
@@ -43,6 +43,6 @@ public class ReviewReplyVO {
 	@Column(name = "REPLY_DATE")
 	private Date replyDate;
 	@Column(name = "REPLY_RATING")
-	private float replyRating;
+	private double replyRating;
 
 }
